test: share coordinator fixture across split tests
