+++ tests/test_coordinator.py
@@ -15,10 +15,9 @@ from custom_components.thessla_green_modbus.modbus_exceptions import (
     ConnectionException,
     ModbusException,
 )
-from custom_components.thessla_green_modbus.registers.loader import (
-    RegisterDef,
-    get_registers_by_function,
-)
+from custom_components.thessla_green_modbus.registers.loader import RegisterDef
+
+from tests.helpers_coordinator import HOLDING_REGISTERS, INPUT_REGISTERS, _make_config_entry
 
 try:
     from homeassistant.const import CONF_HOST, CONF_NAME, CONF_PORT
@@ -31,18 +30,6 @@ except ImportError:
     class UpdateFailed(Exception):  # type: ignore[no-redef]
         pass
 
-
-def _make_config_entry(data: dict, options: dict | None = None) -> MagicMock:
-    """Create a minimal config entry mock."""
-    entry = MagicMock()
-    entry.data = data
-    entry.entry_id = "test"
-    entry.options = options or {}
-    return entry
-
-
-INPUT_REGISTERS = {r.name: r.address for r in get_registers_by_function("04")}
-HOLDING_REGISTERS = {r.name: r.address for r in get_registers_by_function("03")}
 
 # ✅ FIXED: Import correct coordinator class name
 from custom_components.thessla_green_modbus.coordinator import (
@@ -57,30 +44,6 @@ def test_dt_util_timezone_awareness():
     """Ensure coordinator dt util keeps expected callables available."""
     assert callable(coordinator_dt_util.now)
     assert callable(coordinator_dt_util.utcnow)
-
-
-@pytest.fixture
-def coordinator():
-    """Create a test coordinator."""
-    hass = MagicMock()
-    available_registers = {
-        "holding_registers": {"mode", "air_flow_rate_manual", "special_mode"},
-        "input_registers": {"outside_temperature", "supply_temperature"},
-        "coil_registers": {"system_on_off"},
-        "discrete_inputs": set(),
-    }
-    coordinator = ThesslaGreenModbusCoordinator.from_params(
-        hass=hass,
-        host="localhost",
-        port=502,
-        slave_id=1,
-        name="test",
-        scan_interval=30,
-        timeout=10,
-        retry=3,
-    )
-    coordinator.available_registers = available_registers
-    return coordinator
 
 
 def test_coordinator_clamps_effective_batch():
